WIP: Convert to Bootstrap 4
- Remove custom styling in favor of default Bootstrap styles so it is easier for new users to get started.
