+++ startup/client/GlobalStyle.js
@@ -24,19 +24,6 @@ const GlobalStyle = createGlobalStyle`
     --cb-red: #DA5847;
   }
 
-  html {
-    position: relative;
-    min-height: 100%;
-  }
-
-  body {
-    margin: 0;
-    margin-bottom: 80px;
-    padding: 0;
-    font-size: 14px;
-    line-height: 20px;
-  }
-
   body.isViewDocument {
     padding-top: 20px;
   }
@@ -45,25 +32,8 @@ const GlobalStyle = createGlobalStyle`
     display: none;
   }
 
-  form label {
-    display: block;
-  }
-
-  form .control-label {
-    display: block;
-    margin-bottom: 7px;
-  }
-
-  form label.error {
-    display: block;
-    margin-top: 8px;
-    font-size: 13px;
-    font-weight: normal;
-    color: var(--danger);
-  }
-
-  .page-header {
-    margin-top: 0;
+  .tab-pane {
+    padding: 1rem!important;
   }
 
   .table tr td {
@@ -82,6 +52,7 @@ const GlobalStyle = createGlobalStyle`
 
     .page-header {
       margin-top: 20px;
+      margin-bottom: 20px;
     }
   }
 `;
