+++ ui/admin/components/AdminPageHeader/index.js
@@ -1,0 +1,1 @@
+export { default } from './AdminPageHeader';
