+++ ui/admin/components/AdminPageHeader/styles.js
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+const AdminUserHeader = styled.h4`
+  .badge {
+    font-size: 0.875rem;
+  }
+  .badge-Facebook {
+    background: var(--facebook);
+    color: #fff;
+  }
+
+  .badge-Google {
+    background: var(--google);
+    color: #fff;
+  }
+
+  .badge-GitHub {
+    background: var(--github);
+    color: #fff;
+  }
+`;
+
+export default {
+  AdminUserHeader,
+};
