+++ ui/global/components/Navigation/styles.js
@@ -1,15 +1,8 @@
 import styled from 'styled-components';
+import { Navbar } from 'react-bootstrap';
 
-const Navbar = styled.div`
-  background: #f8f9fa !important;
-  border-bottom: 1px solid #e7e7e7;
-  margin-bottom: 20px;
-
-  .navbar-brand a {
-    color: #777;
-  }
-`;
+const StyledNavbar = styled(Navbar)``;
 
 export default {
-  Navbar,
+  StyledNavbar,
 };
