+++ ui/users/pages/Profile/index.js
@@ -1,9 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useMutation, useQuery } from '@apollo/client';
 import { Row, Col, Tabs, Tab } from 'react-bootstrap';
 
 import AccountPageFooter from '../../components/AccountPageFooter';
-import DeleteAccount from '../../components/DeleteAccount';
 import ExportUserData from '../../components/ExportUserData';
 import ProfileForm from '../../components/ProfileForm';
 import UserSettings from '../../components/UserSettings';
@@ -12,7 +11,6 @@ import { updateUser as UPDATE_USER } from '../../mutations/Users.gql';
 import Styles from './styles';
 
 const Profile = () => {
-  const [activeTab, setActiveTab] = useState('profile');
   const { data } = useQuery(GET_USER);
   const [updateUser] = useMutation(UPDATE_USER);
 
@@ -23,21 +21,13 @@ const Profile = () => {
         <h4 className="page-header">
           {thisUser.name ? `${thisUser.name.first} ${thisUser.name.last}` : thisUser.username}
         </h4>
-        <Tabs
-          animation={false}
-          activeKey={activeTab}
-          onSelect={(newTab) => setActiveTab(newTab)}
-          id="admin-user-tabs"
-        >
+        <Tabs>
           <Tab eventKey="profile" title="Profile">
             <Row>
               <Col xs={12} sm={6} md={4}>
                 <ProfileForm user={data} />
                 <AccountPageFooter>
                   <ExportUserData />
-                </AccountPageFooter>
-                <AccountPageFooter>
-                  <DeleteAccount />
                 </AccountPageFooter>
               </Col>
             </Row>
